// src/app/interfaces/multas.interface.ts
// Generated by https://quicktype.io

export interface ResultMultas {
    ok:    boolean;
    msg:   string;
    multa: Multa[];
}
export interface ResultMulta {
    ok:    boolean;
    msg:   string;
    multa: Multa;
}
export interface ResultMultaPost {
    ok:    boolean;
    msg:   string;
    multa: MultaPost;
}

export interface Multa {
    _id:         string;
    titulo:      string;
    descripcion: string;
    resolucion:  string;
    estado:      boolean;
    fecha:       string;
    usuario:     Usuario;
}
export interface MultaPost {
    _id:         string;
    titulo:      string;
    descripcion: string;
    resolucion:  string;
    estado:      boolean;
    fecha:       string;
    usuario:     string;
}

export interface Usuario {
    _id:    string;
    nombre: string;
}